:art: Dienststelleninput into component
